resize icons to 255px max width
Only icons up to 255x255 are supported in DGUSM.  Newer DW_ICON.exe tools
(v7) need to be used to generate files with these large icons though as v5
doesn't work.  256x256 was tried as well since the documentation was
ambiguous as to the max size, but it didn't work.

Only the width was modified; the height was left as-is since it fit.  In
order to fix most icons, a gradient was added so that the icon fades to the
background on the right.  For the progress bar, the entire icon was just
made narrower.