fix numeric input from being accepted
8-byte long ints are not supported.  Use 4-byte long ints instead.  Note
that this created holes in the memory map, but it isn't a big deal.

We can only see 9 digits on the screen anyway, and with 3 decimal digits it
is dangerous to go over 6 integer digits due to the risk of overflow.
Change the integer digits to 6 and decimal digits to 3.  Validation /
conversion can be performed in code for this.