Add missing VP_AUX (ie AUX_PTR / AP) to Bit Variable Icons
2 words (4 bytes) of memory are required for each unique Bit Variable Icon.
These were all being set to 0, which is invalid and overlaps with other VPs
as well as each other.

The DGUS manual describes this as:
   Substitutive variable pointer, reserved 2 words.  User software unable to
   access.

The DGUS II manual describes this as:
   Auxiliary Variable Pointer.  2 words.  User can not use them, should be
   allocated right after VP.

Since placing them immediately after VP would cause the entire memory map to
rearrange, these were just allocated at the end of memory.  That placement
is likely a recommendation rather than a requirement, and it may not even be
applicable for DGUSM.

It is also extremely likely that these words are just used when the display
mode is set to an animated icon, similar to how the "Animation Icon" control
has the second word reserved automatically to store animation status.  In
any case, these are now unique.

Given that "Bit Variable Icons" require 6 bytes of memory compared to just 2
bytes for a standard "Variable Icon", it may be a good idea to use those
instead in the future.